+++ src/components/cards/noticeCard/NoticeCards.stories.ts
@@ -2,6 +2,8 @@ import type { Meta, StoryObj } from "@storybook/web-components-vite";
 import NoticeCard from "./NoticeCard.svelte";
 import { noticeTypes, noticeTypeDefault } from "./NoticeTypes";
 
+import Link from "../../links/link/Link.svelte";
+
 const meta = {
   title: "Cards/NoticeCard",
   component: "khao-card-notice",
@@ -23,10 +25,10 @@ const meta = {
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-export const Default: Story = {
+export const Info: Story = {
   args: {
     type: noticeTypeDefault,
-    title: "A default notice card",
+    title: "An info notice card",
   },
 };
 
@@ -42,4 +44,20 @@ export const Warning: Story = {
     type: "warning",
     title: "A warning notice card",
   },
+};
+
+export const infoExample: Story = {
+  render: () => `<khao-card-notice type="info" title="A primary notice card">
+    Lorem ipsum dolor sit amet, consectetuer <khao-link label="test" iconName="speaker"></khao-link> adipiscing elit. Aenean commodo (<khao-link label="test" iconName="external-link"></khao-link>) eu, pretium quis.Stet clita <khao-link label="test" iconName="cart"></khao-link> kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.</khao-card-notice>`,
+};
+
+export const neutralExample: Story = {
+  render: () => `<khao-card-notice type="neutral" title="A primary notice card">
+    Lorem ipsum dolor sit amet, consectetuer <khao-link label="test" iconName="speaker"></khao-link> adipiscing elit. Aenean commodo (<khao-link label="test" iconName="external-link"></khao-link>) eu, pretium quis.Stet clita <khao-link label="test" iconName="cart"></khao-link> kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.</khao-card-notice>`,
+};
+
+
+export const warningExample: Story = {
+  render: () => `<khao-card-notice type="warning" title="A warning notice card">
+    Lorem ipsum dolor sit amet, consectetuer <khao-link label="test" iconName="speaker"></khao-link> adipiscing elit. Aenean commodo (<khao-link label="test" iconName="external-link"></khao-link>) eu, pretium quis.Stet clita <khao-link label="test" iconName="cart"></khao-link> kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.</khao-card-notice>`,
 };