+++ src/index.ts
@@ -1,13 +1,11 @@
-import type SocialButtons from "./components/widgets/socialButtons/SocialButtons.svelte";
-
 export * as Style from "@der-reiskoch/khao-malet/css/index.css";
 
 export * as Ad from "./components/ads/ad/Ad.svelte";
 export * as Button from "./components/buttons/button/Button.svelte";
 export * as Card from "./components/cards/card/Card.svelte";
 export * as MarkerPin from "./components/map/markerPin/MarkerPin.svelte";
+export * as Link from "./components/links/link/Link.svelte";
 export * as NoticeCard from "./components/cards/noticeCard/NoticeCard.svelte";
 export * as PayPalDonateButton from "./components/buttons/payPalDonateButton/PayPalDonateButton.svelte";
 export * as ShareButtons from "./components/widgets/shareButtons/ShareButtons.svelte";
-export * as SocialButtons from "./components/widgets/socialButtons/SocialButtons.svelte";
 export * as SteadyButton from "./components/buttons/steadyButton/SteadyButton.svelte";